requirements.txt
streamlit
openai
google-generativeai
python-dotenv
PyPDF2
Pillow